+++ src/app/layout.tsx
@@ -1,7 +1,5 @@
 import type { Metadata } from "next";
 import "./globals.css";
-import Navbar from "../components/nav/Navbar";
-import Footer from "../components/footer/Footer";
 import Script from "next/script";
 
 export const metadata: Metadata = {
